+++ src/main/java/open/commons/spring/jdbc/dao/AbstractSingleDataSourceDao.java
@@ -111,7 +111,7 @@ public abstract class AbstractSingleDataSourceDao extends AbstractGenericDao {
             StringBuffer msg = new StringBuffer();
             msg.append("con=");
             msg.append(con.toString());
-            msg.append("con-to-work=");
+            msg.append(", con-to-work=");
             msg.append(conToWork != null ? conToWork.toString() : null);
 
             DataAccessException dae = jdbcTemplate.getExceptionTranslator().translate("ConnectionCallback", msg.toString(), e);